fix: bug in cadidate partition determination
When writing multiple updates, w/o a reboot in-between, the wrong
candidate kernel/rootfs partition could be mounted. This could result in
the old/current kernel being copied to the new/updated partition.

look at `bootcount` when determining candidate partition(s) to mount. This variable will be 0 when an update has been written, but not yet active.

resolves #6