fix: use upgrade_available to determine current rootfs partition
Previously added logic to determine the correct (newest/candidate)
partition to pull kernel from and copy to the secondary kernel
partitons. There was a previous issue where this could be incorrect. I
used the mender env `bootcount` to determine this (if we were or weren't
in the middle of an update).

This hasn't been reliable. Moving to `upgrade_available`, which seems to
work better.